+++ src/main/java/fr/raksrinana/fallingtree/FallingTree.java
@@ -1,9 +1,9 @@
 package fr.raksrinana.fallingtree;
 
 import fr.raksrinana.fallingtree.config.Config;
-import me.shedaniel.clothconfig2.forge.api.ConfigBuilder;
-import net.minecraft.util.text.StringTextComponent;
+import fr.raksrinana.fallingtree.config.cloth.ClothConfigHook;
 import net.minecraftforge.fml.ExtensionPoint;
+import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
@@ -14,17 +14,15 @@ import org.apache.commons.lang3.tuple.Pair;
 public class FallingTree{
 	public static final String MOD_ID = "falling_tree";
 	
-	public FallingTree(){
+	public FallingTree() throws ClassNotFoundException, IllegalAccessException, InstantiationException{
 		ModLoadingContext.get().registerExtensionPoint(ExtensionPoint.DISPLAYTEST, () -> Pair.of(() -> FMLNetworkConstants.IGNORESERVERONLY, (a, b) -> true));
 		ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, Config.COMMON_SPEC);
-		ModLoadingContext.get().registerExtensionPoint(ExtensionPoint.CONFIGGUIFACTORY, () -> (mc, parent) -> {
-			ConfigBuilder builder = ConfigBuilder.create()
-					.setParentScreen(parent)
-					.setTitle(new StringTextComponent("FallingTree"));
-			
-			Config.COMMON.fillConfigScreen(builder);
-			
-			return builder.build();
-		});
+		
+		if(ModList.get().isLoaded("cloth-config")){
+			Class.forName("fr.raksrinana.fallingtree.config.cloth.ClothConfigHook")
+					.asSubclass(ClothConfigHook.class)
+					.newInstance()
+					.load();
+		}
 	}
 }